changes requested by the review
