+++ src/ConfigData/TableConfigurations/TableConfiguration.cs
@@ -149,7 +149,7 @@
             }
         }
 
-        public string ReadFromJSON(string JSONstring, int hwnd)
+        public string ReadFromJSON(string JSONstring)
         {
             string errors = null;
             try
@@ -178,7 +178,6 @@
 
 
                 var addonConfig = new AddOnDataProcessingConfiguration();
-                addonConfig.BelongsToHwnd = hwnd;
                 addonConfig.ReadFromJSON(jsonConfig?.qHyperCubeDef);
                 AddOnData.Clear();
                 AddOnData.Add(addonConfig);
