+++ src/Controls/BusyCircle.xaml.cs
@@ -15,15 +15,15 @@
             InitializeComponent();
         }
 
-        #region Circlethickness - DP       
-        public Thickness Circlethickness
+        #region Circlethickness - DP          
+        public double Circlethickness
         {
-            get { return (Thickness)this.GetValue(CirclethicknessProperty); }
+            get { return (double)this.GetValue(CirclethicknessProperty); }
             set { this.SetValue(CirclethicknessProperty, value); }
         }
 
         public static readonly DependencyProperty CirclethicknessProperty = DependencyProperty.Register(
-         "Circlethickness", typeof(Thickness), typeof(BusyCircle), new FrameworkPropertyMetadata(new Thickness(1), FrameworkPropertyMetadataOptions.BindsTwoWayByDefault));
+         "Circlethickness", typeof(double), typeof(BusyCircle), new FrameworkPropertyMetadata(1.0, FrameworkPropertyMetadataOptions.BindsTwoWayByDefault));
         #endregion
     }
 }
