+++ src/Controls/ControlLoader.xaml.cs
@@ -1,10 +1,11 @@
 ﻿namespace daVinci.Controls
 {
+    using leonardo.AttachedProperties;
     #region Usings
     using NLog;
     using System;
     using System.Windows;
-    using System.Windows.Controls; 
+    using System.Windows.Controls;
     #endregion
     /// <summary>
     /// Interaction logic for ControlLoader.xaml
@@ -20,7 +21,7 @@
         }
 
         public Type TypeToCreate { get; set; }
-
+        public int Hwnd { get; set; }
         private void UserControl_Loaded(object sender, RoutedEventArgs e)
         {
             try
@@ -28,6 +29,7 @@
                 if (TypeToCreate != null)
                 {
                     var control = (Control)Activator.CreateInstance(TypeToCreate);
+                    control.SetValue(ThemeProperties.HwndProperty, Hwnd);
                     control.DataContext = DataContext;
                     Content = control;
                 }
