+++ src/Controls/DimensionColumnDataView.xaml
@@ -54,7 +54,6 @@
     <StackPanel Margin="5,0,5,0">
         <daViniciControls:DavInputbox
             Margin="0,10,0,5"
-            Hwnd="{Binding DimensionMeasure.BelongsToHwnd}"
             IsInputEnabled="True"
             LabelText="{l:Loc Translate_common:Common_Field}"
             RightCommandForeground="{StaticResource Lui.Brush.Orange}"
@@ -62,7 +61,6 @@
             Visibility="{Binding IsExpression, Converter={StaticResource TrueToVisibleConverter}}" />
         <daViniciControls:DavInputbox
             Margin="0,10,0,5"
-            Hwnd="{Binding DimensionMeasure.BelongsToHwnd}"
             IsInputEnabled="True"
             LabelText="{l:Loc Translate_common:Common_Label}"
             Text="{Binding FieldLabel, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
