+++ src/Controls/LoopConfiguration.xaml
@@ -61,7 +61,6 @@
                 IsEnabled="{Binding LoopConfigurationSelected.JsonValid, RelativeSource={RelativeSource FindAncestor, AncestorType=daViniciControls:LoopConfiguration}}" />
 
             <daViniciControls:DavInputbox
-                Hwnd="{Binding LoopConfigurationSelected.BelongsToHwnd, RelativeSource={RelativeSource FindAncestor, AncestorType=daViniciControls:LoopConfiguration}}"
                 IsEnabled="{Binding LoopConfigurationSelected.EnableExpression, RelativeSource={RelativeSource FindAncestor, AncestorType=daViniciControls:LoopConfiguration}}"
                 LabelText="Sheet Name"
                 Text="{Binding LoopConfigurationSelected.SheetNameExpression, RelativeSource={RelativeSource FindAncestor, AncestorType=daViniciControls:LoopConfiguration}, UpdateSourceTrigger=PropertyChanged}" />
