+++ src/Controls/PropertyPanel.xaml.cs
@@ -30,17 +30,6 @@
             InitializeComponent();
         }
 
-        #region OwnerHwndHwnd DP
-        public int OwnerHwnd
-        {
-            get { return (int)this.GetValue(OwnerHwndProperty); }
-            set { this.SetValue(OwnerHwndProperty, value); }
-        }
-
-        public static readonly DependencyProperty OwnerHwndProperty = DependencyProperty.Register(
-         "OwnerHwnd", typeof(int), typeof(PropertyPanel), new FrameworkPropertyMetadata(-1, FrameworkPropertyMetadataOptions.BindsTwoWayByDefault));
-        #endregion
-
         private bool showImport = false;
         public bool ShowImport
         {
