+++ src/Styles/ColumnChooserStylePivot.xaml
@@ -17,10 +17,7 @@
 
     <tableconfigurations:ColumnOrderIndexProvider x:Key="ColumnOrderIndexProvider" />
     <configdata:ColumnOrderComparer x:Key="ColumnOrderComparer" />
-
     <daVinciConverter:ColumnNameConverter x:Key="ColumnNameConverter" />
-
-    <daVinciConverter:ControlCreationConverter x:Key="ControlCreationConverter" />
 
     <daVinciControls:PivotRowTypeColumnFilter x:Key="PivotRowTypeColumnFilter" />
     <daVinciControls:PivotColumnTypeColumnFilter x:Key="PivotColumnTypeColumnFilter" />
