+++ src/Styles/ColumnConfigurationStyle.xaml
@@ -1,6 +1,7 @@
 ﻿<ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:LuiAttached="clr-namespace:leonardo.AttachedProperties;assembly=leonardo-wpf"
     xmlns:LuiControls="clr-namespace:leonardo.Controls;assembly=leonardo-wpf"
     xmlns:LuiConverter="clr-namespace:leonardo.Converter;assembly=leonardo-wpf"
     xmlns:daVinciConfig="clr-namespace:daVinci.ConfigData"
@@ -31,7 +32,14 @@
             </Setter.Value>
         </Setter>
         <Setter Property="LuiControls:LuiAccordionItem.Index" Value="{Binding SortCriterias.ColumnOrderIndex, Mode=TwoWay}" />
-        <Setter Property="LuiControls:LuiAccordionItem.Content" Value="{Binding Path=., Converter={StaticResource ControlCreationConverter}, ConverterParameter={x:Type daVinciControls:DimensionColumnDataView}}" />
+        <Setter Property="LuiControls:LuiAccordionItem.Content">
+            <Setter.Value>
+                <MultiBinding Converter="{StaticResource ControlCreationConverter}" ConverterParameter="{x:Type daVinciControls:DimensionColumnDataView}">
+                    <Binding Path="." />
+                    <Binding Path="(LuiAttached:ThemeProperties.Hwnd)" RelativeSource="{RelativeSource Self}" />
+                </MultiBinding>
+            </Setter.Value>
+        </Setter>
     </Style>
 
 
@@ -50,7 +58,14 @@
             </Setter.Value>
         </Setter>
         <Setter Property="LuiControls:LuiAccordionItem.Index" Value="{Binding SortCriterias.ColumnOrderIndex, Mode=TwoWay}" />
-        <Setter Property="LuiControls:LuiAccordionItem.Content" Value="{Binding Path=., Converter={StaticResource ControlCreationConverter}, ConverterParameter={x:Type daVinciControls:MeasureColumnDataView}}" />
+        <Setter Property="LuiControls:LuiAccordionItem.Content">
+            <Setter.Value>
+                <MultiBinding Converter="{StaticResource ControlCreationConverter}" ConverterParameter="{x:Type daVinciControls:MeasureColumnDataView}">
+                    <Binding Path="." />
+                    <Binding Path="(LuiAttached:ThemeProperties.Hwnd)" RelativeSource="{RelativeSource Self}" />
+                </MultiBinding>
+            </Setter.Value>
+        </Setter>
     </Style>
 
 
@@ -80,7 +95,14 @@
             </Setter.Value>
         </Setter>
         <Setter Property="LuiControls:LuiAccordionItem.Index" Value="{Binding SortCriterias.SortOrderIndex, Mode=TwoWay}" />
-        <Setter Property="LuiControls:LuiAccordionItem.Content" Value="{Binding Path=., Converter={StaticResource ControlCreationConverter}, ConverterParameter={x:Type daVinciControls:SortSetting}}" />
+        <Setter Property="LuiControls:LuiAccordionItem.Content">
+            <Setter.Value>
+                <MultiBinding Converter="{StaticResource ControlCreationConverter}" ConverterParameter="{x:Type daVinciControls:SortSetting}">
+                    <Binding Path="." />
+                    <Binding Path="(LuiAttached:ThemeProperties.Hwnd)" RelativeSource="{RelativeSource Self}" />
+                </MultiBinding>
+            </Setter.Value>
+        </Setter>
     </Style>
